+++ changelog.md
@@ -1,2 +1,6 @@
+# 1.0.1
+- Fixed improper error handling if user doesn't exist
+- Fixed Android issue where levels page would not load correctly from user searches
+
 # 1.0.0
 - Release